♻️ Isolate common methods that send error messages
