+++ src/main/java/com/und/server/security/CustomAuthenticationEntryPoint.java
@@ -2,12 +2,10 @@ package com.und.server.security;
 
 import java.io.IOException;
 
-import org.springframework.http.MediaType;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.und.server.exception.ServerErrorResult;
 
 import jakarta.servlet.http.HttpServletRequest;
@@ -18,7 +16,7 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint {
 
-	private final ObjectMapper objectMapper;
+	private final SecurityErrorResponseWriter errorResponseWriter;
 
 	@Override
 	public void commence(
@@ -26,15 +24,6 @@ public class CustomAuthenticationEntryPoint implements AuthenticationEntryPoint 
 		HttpServletResponse response,
 		AuthenticationException authException
 	) throws IOException {
-		ServerErrorResult errorResult = ServerErrorResult.UNAUTHORIZED_ACCESS;
-		response.setStatus(errorResult.getHttpStatus().value());
-		response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-		response.setCharacterEncoding("UTF-8");
-		response.getWriter().write(objectMapper.writeValueAsString(
-			new ErrorResponse(errorResult.name(),
-			errorResult.getMessage())
-		));
+		errorResponseWriter.sendErrorResponse(response, ServerErrorResult.UNAUTHORIZED_ACCESS);
 	}
-
-	private record ErrorResponse(String code, Object message) { }
 }
